Added new dataset and code to generate it.
The hyg dataset had column names that were really unintuitive and over 1600 observations that we cannot use.  I renamed the column names and removed the extra observations.  In the jupyter notebook, I also included a description of the attributes that can be used in a codebook and provide the link to the site where I obtained that information.